# docs/openapi.yml
openapi: 3.0.0
servers:
  - description: Production
    url: http://localhost:9090
info:
  description: Codewind API
  version: 0.1.0
  title: Codewind API
  contact:
    name: API support
    email: example@ibm.com
  license:
    name: Codewind Developer License
    url: 'https://github.com/eclipse/codewind/blob/master/LICENSE/LICENSE.md'

# tags:
  # - name: users
  #   description: Operations available to regular users
  # - name: admin
  #   description: Operations available only to authenticated admins
paths:
...
  /api/v1/templates:
    get:
      summary: List all templates Codewind is aware of
      parameters:
        - name: projectStyle
          in: query
          schema:
            type: string
          required: false
          description: Specify this to filter templates returned by project style
        - name: showEnabledOnly
          in: query
          schema:
            type: string
            enum: ['true', 'false']
          required: false
          description: Set this to 'true' to return only templates from enabled repositories
      responses:
        200:
          description: Returns all the templates that Codewind can generate a project from
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
        204:
          description: No templates found
  /api/v1/templates/repositories:
    get:
      summary: List all template repositories Codewind is aware of
      responses:
        200:
          description: Returns all the templates repositories that Codewind will use to find templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateRepo'
    post:
      summary: Make Codewind aware of a template repository
      requestBody:
        description: repository details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateRepo'
      responses:
        200:
          description: Returns all the templates repositories that Codewind will use to find templates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    url:
                      type: string
                    description:
                      type: string
        400:
          description: Bad request
    delete:
      summary: Remove a template repository from the list Codewind is aware of
      requestBody:
        description: repository details
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                    type: string
      responses:
        200:
          description: Returns all the templates repositories that Codewind will use to find templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateRepo'
        400:
          description: Bad request
  /api/v1/batch/templates/repositories:
    patch:
      summary: Batch change settings for template repositories of which Codewind is aware
      requestBody:
        description: repository settings
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/TemplateRepoSetting'
      responses:
        207:
          description: Returns success/failure statuses of settings operations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - status
                    - requestedOperation
                  properties:
                    status:
                      type: integer
                    requestedOperation:
                      $ref: '#/components/schemas/TemplateRepoSetting'
                    error:
                      type: string
  /api/v1/batch/templates/extensions:
    patch:
      summary: Batch change settings for template extensions of which Codewind is aware
      requestBody:
        description: extension settings
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/TemplateExtensionSetting'
      responses:
        207:
          description: Returns success/failure statuses of settings operations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - status
                    - requestedOperation
                  properties:
                    status:
                      type: integer
                    requestedOperation:
                      $ref: '#/components/schemas/TemplateExtensionSetting'
                    error:
                      type: string
...
components:
  schemas:
...
    Template:
      type: object
      required:
        - label
        - description
        - language
        - url
        - projectType
      properties:
        label:
          type: string
          example: "python hello world"
        description:
          type: string
          example: "python microservice"
        language:
          type: string
          example: "python"
        url:
          type: string
          example: "https://github.com/microclimate-dev2ops/SVTPythonTemplate"
        projectType:
          type: string
          example: "docker"
    TemplateRepo:
      type: object
      required:
        - url
        - description
      properties:
        url:
          type: string
        description:
          type: string
        enabled:
          type: boolean
...
    TemplateRepoSetting:
      type: object
      required:
        - op
        - url
        - value
      properties:
        op:
          description: repository setting to change
          type: string
          enum: [enable]
        url:
          description: url of the modified template repository
          type: string
          example: https://raw.githubusercontent.com/kabanero-io/codewind-templates/master/devfiles/index.json
        value:
          type: string
          example: 'true'
    TemplateExtensionSetting:
      type: object
      required:
        - op
        - name
        - value
      properties:
        op:
          description: extension setting to change
          type: string
          enum: [enable]
        name:
          description: name of the modified template extension
          type: string
          example: AppsodyExtension
        value:
          type: string
          example: 'true'
...